requirements.txt
streamlit
st-pages
langchain
openai
tiktoken
transformers
python-docx
pandas
numpy
python-dotenv